gitignore only .env.*local instead of all env
create-react-app ships with dotenv and can automatically wire in env
vars based on NODE_ENV. In web3, it may be common to git commit env vars
like an ethrpc endpoint, so we only want gitignore to override the local
env files, which, as the 'local' name implies, are intended strictly for
local development.

However, we won't use create-react-app's dotenv support to facilitate
our environment switching. This is because we want the ability to build
a regular production build, but with dependencies set to test, so that
users can try a prod build of the site for a demo or testing purposes.
(eg. this would be a canonical prod-test environment.)
Unfortunately, create-react-app does not support this feature.
References:
- create-react-app's dotenv loader code: https://github.com/facebook/create-react-app/blob/bb64e31a81eb12d688c14713dce812143688750a/packages/react-scripts/config/env.js
- create-react-app's docs for dotenv https://create-react-app.dev/docs/adding-custom-environment-variables/
- an issue tracking the feature request to inject the dotenv env file
  into the create-react-app build https://github.com/facebook/create-react-app/issues/11335