Add usedappCurrencies.ts
usedappCurrencies.ts is a list of definitions for
@usedapp/core.NativeCurrency and @usedapp/core.Token for popular tokens
on popular networks and testnets.

Currently we support most of the cross product of (ETH, DAI, USDC) X
(mainnet, Kovan, Optimism, OptimismKovan, Arbitrum, ArbitrumRinkeby).

These definitions are rolled up into the exported `tokens` which
changes based on whether or not the app is built in production mode.
For example, DAI on OptimismKovan is excluded from tokens when
isProduction==false.