Add observer.ts
Observer/ObservableValue enables subscriptions to synchronously observe
a stream of values of type T. ObservableValue is memoryless; only the
current (latest) value is retained.

The primary use case for ObservableValue is to avoid unwanted re-renders
of React components because a React component re-renders when its props
or state are updated.