Add connectedWalletContextProvider.tsx
Adds ConnectedWalletAddressContextObserverProvider, a state container
intended to be used in the root of the app to keep track of the
connected wallet address and map it to a rich context type
AddressContext.

Clients use the useConnectedWalletAddressContext hook which
automagically provides an AddressContext for the connected wallet, which
will soon include live token balances. Crucially, this hook allows React
components anywhere in the tree to subscribe to AddressContext without
needing to pass any props down and without an ancestor component needing
to track the AddressContext or its inputs as state. This allows
real-time tracking of AddressContext data without triggering re-renders
of ancestor/intermediate components.