Add serialization tech demo
Adds a non-production demo primarily intended to measure the efficacy
of a proposed serialization pipeline.

Our serialized values are intended to be stored in URLs and so our goal
is to produce URLs that are as short as possible.

Given a Checkout object, we serialize it, gzip it, and base64 it, and
then reverse the process and ensure the deserialized result is correct.

For serialization, we test two competing libraries, @msgpack/msgpack vs.
@shelacek/ubjson. We find that @msgpack/msgpack produces a slightly
smaller result on our single test object.

For compression, a library survey found only a single decent, modern,
fully-js compresion library: fflate. fflate contains a few different
compression algorithms and we try only gzip.

For base64, we use our modified base64 algorithm that produces values
that may be included in an URL without encoding.

In our experiment, the final base64 length of our compressed, serialized
test Checkout was 288 chars for msgpack and 296 chars for ubjson.

Both msgpack and ubjson are schemaless serializers that don't require us
to write and maintain a separate serialization schema. However, we
expect that a future transition to a schema-based serializer would
likely yield much shorter final URL sizes. But, we want to start simple,
and writing and maintaining a serialization schema is quite expensive.