Add switchNetwork
switchNetwork is our temporary implementation of switchNetwork
based loosely on usedapp and Uniswap's implementations.

Our switchNetwork is a temporary lib because we'll soon migrate to wagmi
which has their own implementation of switchNetwork.

The hack in our switchNetwork is that it contains its own hardcoded
network definitions needed to auto-add a network the user's wallet
doesn't have yet. Wagmi's switchNetwork uses their network definitions,
which are much more complete and better maintained, and so we don't want
to maintain them as well.